netaxe/requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.6.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
channels==2.4.0
channels-redis==3.4.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==37.0.4
daphne==2.5.0
Deprecated==1.2.13
Django==3.2
django-annoying==0.10.6
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-filter==22.1
django-guardian==2.4.0
django-jsonstore==0.5.0
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-ratelimit==3.0.1
django-redis==5.2.0
django-simple-captcha==0.5.17
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-extensions==0.7.1
gitdb==4.0.9
GitPython==3.1.3
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
kombu==5.2.4
lizard==1.17.10
msgpack==1.0.4
mypy-extensions==0.4.3
PyMySQL==1.0.2
nornir==3.3.0
nornir-jinja2==0.2.0
nornir-napalm==0.2.0
nornir-netmiko==0.1.2
nornir-utils==0.2.0
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.30
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDriller==2.1
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4
redis==4.3.4
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
smmap==5.0.0
sqlparse==0.4.2
ttp==0.6.0
Twisted==22.4.0
txaio==22.2.1
types-pytz==2022.1.2
typing_extensions==4.2.0
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.0
zope.interface==5.4.0
pymongo==3.10.1
# rest api 日志记录模块
drf-api-tracking==1.8.0
xmltodict==0.13.0
requests==2.28.1
pycryptodome==3.15.0
gitdb2==4.0.2
GitPython==3.1.3
supervisor==4.2.4
flower==1.2.0
coreapi==2.3.3